Moved Error component to UI components directory
