+++ src/components/sections/Experiences/Experiences.tsx
@@ -3,7 +3,7 @@
 import { Section } from '@/components/ui/Section';
 import { PortfolioItems } from '@/components/ui/PortfolioItems';
 import { Skeletons } from '@/components/ui/Skeletons';
-import { Error } from '@/components/Error';
+import { Error } from '@/components/ui/Error';
 import type { PortfolioItem as PortfolioItemType } from '@/lib/types';
 
 type ExperiencesProps = {
