+++ src/components/ui/Error/Error.module.scss
